move affinity logic to new file
